Scrape Tweaks
Scraper now looks at everything collected and filters through the results instead of just taking the first x posts. Posts are now used if they have descriptions more than n characters and if the post image is of landscape perspective.

Main script now has a longer maximum execution time due to the number of steps involved in the installation process.

Hashtags are parsed from descriptions and used to generated a hashtag cloud on the left panel.